webapp/requirements.dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -c requirements.txt --extra dev --generate-hashes pyproject.toml
asgiref==3.8.1 \
    --hash=sha256:3e1e3ecc849832fe52ccf2cb6686b7a55f82bb1d6aee72a58826471390335e47 \
    --hash=sha256:c343bd80a0bec947a9860adb4c432ffa7db769836c64238fc34bdc3fec84d590
    # via
    #   -c requirements.txt
    #   django
django==5.1 \
    --hash=sha256:848a5980e8efb76eea70872fb0e4bc5e371619c70fffbe48e3e1b50b2c09455d \
    --hash=sha256:d3b811bf5371a26def053d7ee42a9df1267ef7622323fe70a601936725aa4557
    # via
    #   -c requirements.txt
    #   orderly-ape (pyproject.toml)
    #   django-autocomplete-light
    #   django-debug-toolbar
    #   django-fsm-2
    #   djangorestframework
django-autocomplete-light==3.11.0 \
    --hash=sha256:212576a17e3308ef7ca77e280b86684167916d2091d4b73640f38845d9516328
    # via
    #   -c requirements.txt
    #   orderly-ape (pyproject.toml)
django-debug-toolbar==4.4.6 \
    --hash=sha256:36e421cb908c2f0675e07f9f41e3d1d8618dc386392ec82d23bcfcd5d29c7044 \
    --hash=sha256:3beb671c9ec44ffb817fad2780667f172bd1c067dbcabad6268ce39a81335f45
    # via orderly-ape (pyproject.toml)
django-environ==0.11.2 \
    --hash=sha256:0ff95ab4344bfeff693836aa978e6840abef2e2f1145adff7735892711590c05 \
    --hash=sha256:f32a87aa0899894c27d4e1776fa6b477e8164ed7f6b3e410a62a6d72caaf64be
    # via
    #   -c requirements.txt
    #   orderly-ape (pyproject.toml)
django-fsm-2==3.0.0 \
    --hash=sha256:19887dcd6009d9b785f6114527e6542f380c5af7061c47bebefcbad9ace730aa \
    --hash=sha256:5ca98a4c1f5537a54433f3963cd8473dcdc039d9bc8c8045250d290f84a25b35
    # via
    #   -c requirements.txt
    #   orderly-ape (pyproject.toml)
django-types==0.19.1 \
    --hash=sha256:5ae7988612cf6fbc357b018bbc3b3a878b65e04275cc46e0d35d66a708daff12 \
    --hash=sha256:b3f529de17f6374d41ca67232aa01330c531bbbaa3ac4097896f31ac33c96c30
    # via orderly-ape (pyproject.toml)
djangorestframework==3.15.2 \
    --hash=sha256:2b8871b062ba1aefc2de01f773875441a961fefbf79f5eed1e32b2f096944b20 \
    --hash=sha256:36fe88cd2d6c6bec23dca9804bab2ba5517a8bb9d8f47ebc68981b56840107ad
    # via
    #   -c requirements.txt
    #   orderly-ape (pyproject.toml)
djangorestframework-types==0.8.0 \
    --hash=sha256:5136212ec0df2c11d5ebcf217246f606e68f1d46a770419de84ed676bdab0113 \
    --hash=sha256:bed5dc8ce4f3500a4d29632e333262059e2218f2291c88313ce8443c24e50a76
    # via orderly-ape (pyproject.toml)
sqlparse==0.5.1 \
    --hash=sha256:773dcbf9a5ab44a090f3441e2180efe2560220203dc2f8c0b0fa141e18b505e4 \
    --hash=sha256:bb6b4df465655ef332548e24f08e205afc81b9ab86cb1c45657a7ff173a3a00e
    # via
    #   -c requirements.txt
    #   django
    #   django-debug-toolbar
types-psycopg2==2.9.21.20240819 \
    --hash=sha256:4ed6b47464d6374fa64e5e3b234cea0f710e72123a4596d67ab50b7415a84666 \
    --hash=sha256:c9192311c27d7ad561eef705f1b2df1074f2cdcf445a98a6a2fcaaaad43278cf
    # via django-types
whitenoise==6.7.0 \
    --hash=sha256:58c7a6cd811e275a6c91af22e96e87da0b1109e9a53bb7464116ef4c963bf636 \
    --hash=sha256:a1ae85e01fdc9815d12fa33f17765bc132ed2c54fa76daf9e39e879dd93566f6
    # via
    #   -c requirements.txt
    #   orderly-ape (pyproject.toml)
